00._data_format_translation_servers/python/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["Alexander <alex030tstud.kea.dk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
xmltodict = "^0.13.0"
pyyaml = "^6.0"
pandas = "^1.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
